JAVA-6099 address PR comments
